+++ CHANGELOG.md
@@ -17,7 +17,7 @@ signal-utils 0.3.0 (minor)
 
 #### :rocket: Enhancement
 * `signal-utils`
-  * [#9](https://github.com/NullVoxPopuli/signal-utils/pull/9) New Util: ReactiveArray ([@NullVoxPopuli](https://github.com/NullVoxPopuli))
+  * [#9](https://github.com/NullVoxPopuli/signal-utils/pull/9) New Util: SignalArray ([@NullVoxPopuli](https://github.com/NullVoxPopuli))
 
 #### :house: Internal
 * `signal-utils`
@@ -32,7 +32,7 @@ signal-utils 0.2.0 (minor)
 
 #### :rocket: Enhancement
 * `signal-utils`
-  * [#5](https://github.com/NullVoxPopuli/signal-utils/pull/5) New util: `ReactiveObject` ([@NullVoxPopuli](https://github.com/NullVoxPopuli))
+  * [#5](https://github.com/NullVoxPopuli/signal-utils/pull/5) New util: `SignalObject` ([@NullVoxPopuli](https://github.com/NullVoxPopuli))
 
 #### Committers: 1
 - [@NullVoxPopuli](https://github.com/NullVoxPopuli)